feat: 更新 refresh token 邏輯
如果 refresh token 過期就回傳 token 過期訊息，反之會生成新的 access token 和 refresh token，並且會更新資料庫中的 refresh token
資料，並清理屬於該使用者的過期 token

+++ app/crud/base.py
@@ -79,5 +79,6 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
 
         obj = db.get(self.model, id)
 
-        db.delete(obj)
-        db.commit()
+        if obj:
+            db.delete(obj)
+            db.commit()
